Cleanup modules, add docs
