+++ hal/src/async_hal/irqs_samd11.rs
@@ -1,0 +1,6 @@
+use crate::pac::Interrupt as InterruptEnum;
+use embassy::interrupt::declare;
+
+declare!(DMAC);
+
+declare!(TC1);
