+++ hal/src/async_hal/irqs_samd21.rs
@@ -2,6 +2,7 @@ use crate::pac::Interrupt as InterruptEnum;
 use embassy::interrupt::declare;
 
 declare!(DMAC);
+
 declare!(TC3);
 declare!(TC4);
 declare!(TC5);
