+++ hal/src/thumbv6m/timer.rs
@@ -185,134 +185,3 @@ tc! {
     TimerCounter4: (TC4, tc4_, Tc4Tc5Clock),
     TimerCounter5: (TC5, tc5_, Tc4Tc5Clock),
 }
-
-#[cfg(feature = "async")]
-pub mod async_timer {
-
-    use super::*;
-    use atomic_polyfill::AtomicBool;
-    use core::{
-        sync::atomic::Ordering,
-        task::{Poll, Waker},
-    };
-    use embassy::{interrupt::InterruptExt, waitqueue::AtomicWaker};
-    use futures::future::poll_fn;
-
-    pub trait AsyncCount16: Count16 {
-        type Interrupt: ::embassy::interrupt::InterruptExt;
-        const STATE_ID: usize;
-        unsafe fn on_interrupt(_: *mut ());
-    }
-
-    macro_rules! impl_async_count16 {
-        ($(($TC: ident, $id: expr)),+) => {
-            $(
-                impl AsyncCount16 for $TC {
-                    type Interrupt = crate::interrupt::$TC;
-                    const STATE_ID: usize = $id;
-
-                    unsafe fn on_interrupt(_: *mut ()) {
-                        let tc = crate::pac::Peripherals::steal().$TC;
-                        let intflag = &tc.count_16().intflag;
-
-                        if intflag.read().ovf().bit_is_set() {
-                            // Clear the flag
-                            intflag.modify(|_, w| w.ovf().set_bit());
-                            STATE[Self::STATE_ID].wake();
-                        }
-                    }
-                }
-            )+
-        };
-    }
-
-    #[cfg(feature = "samd11")]
-    impl_async_count16!((TC1, 0));
-
-    #[cfg(feature = "samd21")]
-    impl_async_count16!((TC3, 0), (TC4, 1), (TC5, 2));
-
-    impl<TC: AsyncCount16> TimerCounter<TC> {
-        pub fn into_future<'a, 'self_mut: 'a>(
-            &'self_mut mut self,
-            irq: &'a mut TC::Interrupt,
-        ) -> AsyncTimer<'a, TC> {
-            AsyncTimer::new(self, irq)
-        }
-    }
-
-    // TODO instead of tracking the state manually, we could use ONESHOT
-    // mode and check the STATUS.STOP bit
-    struct State {
-        waker: AtomicWaker,
-        ready: AtomicBool,
-    }
-
-    impl State {
-        const fn new() -> Self {
-            Self {
-                waker: AtomicWaker::new(),
-                ready: AtomicBool::new(false),
-            }
-        }
-
-        fn register(&self, waker: &Waker) {
-            self.waker.register(waker)
-        }
-
-        fn wake(&self) {
-            self.ready.store(true, Ordering::SeqCst);
-            self.waker.wake()
-        }
-
-        fn ready(&self) -> bool {
-            self.ready.swap(false, Ordering::SeqCst)
-        }
-    }
-
-    const STATE_NEW: State = State::new();
-    static STATE: [State; 3] = [STATE_NEW; 3];
-
-    pub struct AsyncTimer<'a, TC>
-    where
-        TC: AsyncCount16,
-    {
-        timer: &'a mut TimerCounter<TC>,
-        irq: &'a mut TC::Interrupt,
-    }
-
-    impl<'a, TC> AsyncTimer<'a, TC>
-    where
-        TC: AsyncCount16,
-    {
-        pub fn new(timer: &'a mut TimerCounter<TC>, irq: &'a mut TC::Interrupt) -> Self {
-            irq.set_handler(TC::on_interrupt);
-            irq.enable();
-            timer.enable_interrupt();
-
-            Self { timer, irq }
-        }
-
-        pub async fn delay_ms(&mut self, count: impl Into<Nanoseconds>) {
-            self.timer.start(count);
-            self.timer.enable_interrupt();
-
-            poll_fn(|cx| {
-                STATE[TC::STATE_ID].register(cx.waker());
-                if STATE[TC::STATE_ID].ready() {
-                    return Poll::Ready(());
-                }
-
-                Poll::Pending
-            })
-            .await;
-        }
-    }
-
-    impl<'a, TC: AsyncCount16> Drop for AsyncTimer<'a, TC> {
-        fn drop(&mut self) {
-            self.irq.remove_handler();
-            self.irq.disable();
-        }
-    }
-}
